Persist filtering when navigating pages
When moving from the main aggregation page to a specific aggregation
page, the filters are kept and also now updated in the left hand column.
Unfortunately, these filters are lost when returning to the main page,
or going to an individual item page, but for now this is probably ok.